Re-enable support for charsets in comtabs.
This was initially taken out in my rewrite of command-tables.

I'm not sure if we want to keep this around, but I am adding it back in
to get modes that use it to work. Later we might want to take this out if
we decide to re-architect the keystroke system further.